Change the behaviour for updating only the lower or upper bounds
When now updating the lower bound the present upper bound is kept
if possible, it's only removed if the new lower bound is >= the
current upper bound.

If only the upper bound is updated, then no lower bound is added,
only a present lower bound is kept.

Thanks to Benno Fünfstück for these changes.